requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.3
Flask-CORS==3.0.10
pyjwt==2.8.0
Flask-Limiter==3.4.1
email_validator==2.1.1
Flask-Bcrypt==1.0.1
